moving custom reductions into separate module
